float -> double

+++ IndividualProject.ino
@@ -5,13 +5,13 @@ const int VOLTAGE_PIN = 34;
 const int WATER_BUTTON_PIN = 35;
 const int ELECTROLYTE_BUTTON_PIN = 32;
 
-const float SOIL_VOLUME = 100.0; // ml
+const double SOIL_VOLUME = 100.0; // ml
 
 // pH dedicated
-const float Ka = 100; // H2SO4
-const float MolarMass = 98.079; // g/mol, H2SO4
+const double Ka = 100; // H2SO4
+const double MolarMass = 98.079; // g/mol, H2SO4
 
-float get_pH(float Hplus)
+double get_pH(double Hplus)
 {
 	return -log10(Hplus);
 }
@@ -23,10 +23,10 @@ GButton electrolyteButton(ELECTROLYTE_BUTTON_PIN);
 
 uint32_t tmr = 0;
 
-float currentWaterVolume = 100;
-float lastVoltage, lastWaterVolume = 100;
+double currentWaterVolume = 100;
+double lastVoltage, lastWaterVolume = 100;
 
-float Hplus = pow(10, -7);
+double Hplus = pow(10, -7);
 
 void setup() {  
 
@@ -39,7 +39,7 @@ void setup() {
 }
 
 void loop() {
-	float currentVoltage = voltageSensor.getInvertedVoltage();
+	double currentVoltage = voltageSensor.getInvertedVoltage();
 
 	if (waterButton.isClick())
 	{
@@ -50,16 +50,16 @@ void loop() {
 
 	currentWaterVolume = currentVoltage / lastVoltage * lastWaterVolume;
 
-	float soilMosture = currentWaterVolume / SOIL_VOLUME * 100.0;
+	double soilMosture = currentWaterVolume / SOIL_VOLUME * 100.0;
 
 	// pH dedicated
 	if (electrolyteButton.isClick())
 	{
-		float electrolyteMass = 10; // g
-		float electrolyteAmount = electrolyteMass / MolarMass; // mol
-		float electrolyteConcentration = electrolyteAmount / 1; // mol/ml // change 1 to addedWaterVolume
+		double electrolyteMass = 10; // g
+		double electrolyteAmount = electrolyteMass / MolarMass; // mol
+		double electrolyteConcentration = electrolyteAmount / 1; // mol/ml // change 1 to addedWaterVolume
 
-		float deltaHplus = -Ka / 2 + sqrt((Ka / 2) * (Ka / 2) + Ka * electrolyteConcentration);
+		double deltaHplus = -Ka / 2 + sqrt((Ka / 2) * (Ka / 2) + Ka * electrolyteConcentration);
 
 		Hplus += deltaHplus;
 	}
